* added x86_64 libs and dlls for cppunit * removed warnings uncovered by switching to 64 bit target, mainly involving a different-sized size_t * decreased warning level on pcre project

+++ KernelSML/src/sml_KernelHelpers.cpp
@@ -210,7 +210,7 @@ char *symbolToString (Symbol *sym,
 			return dest;
 		}
 		determine_possible_symbol_types_for_string (sym->sc.name,
-			(int)strlen (sym->sc.name),
+			strlen (sym->sc.name),
 			&possible_id,
 			&possible_var,
 			&possible_sc,
